Convert from int to long

+++ timing-tests/vEB-tree-wrapper.cc
@@ -1,8 +1,8 @@
 #include "vEB-tree-wrapper.h"
 using namespace std;
 
-VebTreeWrapper::VebTreeWrapper(const std::vector<double>& weights) : tree(std::vector<int>(2)) {
-	std::vector<int> v;
+VebTreeWrapper::VebTreeWrapper(const std::vector<double>& weights) : tree(std::vector<long>(2)) {
+	std::vector<long> v;
 	for (size_t i = 0; i < weights.size(); i++) {
 		v.push_back(i);
 	}
@@ -13,7 +13,7 @@ VebTreeWrapper::~VebTreeWrapper() {
 	// noop
 }
 
-bool VebTreeWrapper::contains(int key) const {
+bool VebTreeWrapper::contains(long key) const {
 	return tree.contains(key);
 }
